+++ src/StarisApi/Attributes/NotProcessImageAttribute.cs
@@ -1,4 +1,4 @@
-﻿namespace StarisApi.Models;
+﻿namespace StarisApi.Attributes;
 
 [AttributeUsage(AttributeTargets.Class, Inherited = false)]
 public sealed class NotProcessImageAttribute : Attribute
