+++ src/StarisApi/Models/Movies/Movie.cs
@@ -1,4 +1,5 @@
-﻿using StarisApi.Dtos;
+﻿using StarisApi.Attributes;
+using StarisApi.Dtos;
 using StarisApi.Handlers;
 using StarisApi.Models.CharactersMovies;
 using StarisApi.Models.MoviesPlanet;
